fix(engineering-analytics): move job-logs worker into canonical logic/ dir
engineering_analytics is an isolated product; its import-linter only enforces
the canonical backend dirs (presentation, facade, logic, models). The job-logs
worker lived at backend/job_logs/, which the product structure lint flagged as
unrecognized, and its tests werent covered by the backend:test pytest paths so
they would never run in CI.

Move the package to backend/logic/job_logs/ (preserving history) and repoint the
internal cross-imports. Add the new test dir to the backend:test paths so the
tests are actually collected.

Also fix the mypy errors the worker tripped (structural only, runtime behavior
unchanged):
- emitter: build the exporter before setting _enabled and guard the processor
  add with is not None, so mypy narrows LogExporter | None to LogExporter.
- emitter: broaden emit_log_archive attributes to Mapping[str, object]; it
  already filters to scalars internally, so accepting object values is honest
  and matches the test call sites.
