+++ products/engineering_analytics/package.json
@@ -1,7 +1,7 @@
 {
     "name": "@posthog/products-engineering-analytics",
     "scripts": {
-        "backend:test": "pytest -c ../../pytest.ini --rootdir ../.. backend/tests -v --tb=short",
+        "backend:test": "pytest -c ../../pytest.ini --rootdir ../.. backend/tests backend/logic/job_logs/test -v --tb=short",
         "backend:contract-check": "echo 'Contract files unchanged'"
     },
     "dependencies": {
